% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_start_stop_table.R
\name{build_start_stop_table}
\alias{build_start_stop_table}
\title{Transforming data into start - stop data}
\usage{
build_start_stop_table(data)
}
\arguments{
\item{data}{}
}
\value{
the table transform into a start stop table
}
\description{
Transforming data into start - stop data
}
