% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildOnoTable.R
\name{buildOnoTable}
\alias{buildOnoTable}
\title{Transform data into a table of occurrences - non occurrences (ono)}
\usage{
buildOnoTable(df)
}
\arguments{
\item{df}{a start/stop table created with the function buildStartStopTable
(with columns : start, end, observation, subject, behavior}
}
\value{
returns a table of occurences - non occurences.
}
\description{
Transform data into a table of occurrences - non occurrences (ono)
}
