% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_observations_labels.R
\name{get_all_observations_labels}
\alias{get_all_observations_labels}
\title{finding all observations to create a vector}
\usage{
get_all_observations_labels(data)
}
\arguments{
\item{data}{a data table normalized by the build_start_stop_table function}
}
\value{
a vector which contain all the observations observed
}
\description{
finding all observations to create a vector
}
