% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_subjects_labels.R
\name{get_all_subjects_labels}
\alias{get_all_subjects_labels}
\title{finding all subjects to create a vector}
\usage{
get_all_subjects_labels(data, observation_column)
}
\arguments{
\item{data}{a data table normalized by the build_start_stop_table function}
}
\value{
a vector which contain all the subjects observed
}
\description{
finding all subjects to create a vector
}
