% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_table.R
\name{get_data_table}
\alias{get_data_table}
\title{Get the data table from a file}
\usage{
get_data_table(inFile, header, sep, quote)
}
\arguments{
\item{inFile}{}

\item{header}{}

\item{sep}{}

\item{quote}{}
}
\value{
the table table imported
}
\description{
Get the data table from a file
}
