% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBehaviorsLabels.R
\name{getBehaviorsLabels}
\alias{getBehaviorsLabels}
\title{finding all behaviors to create a vector}
\usage{
getBehaviorsLabels(data, observation_column, subject_column)
}
\arguments{
\item{data}{a data table normalized by the build_start_stop_table function}

\item{observation_column}{name of the observation column}

\item{subject_column}{name of the subject column}
}
\value{
a vector which contain all the behaviors observed
}
\description{
finding all behaviors to create a vector
}
