% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getObservationsLabels.R
\name{getObservationsLabels}
\alias{getObservationsLabels}
\title{finding all observations to create a vector}
\usage{
getObservationsLabels(data)
}
\arguments{
\item{data}{a data table normalized by the build_start_stop_table function}
}
\value{
a vector which contain all the observations observed
}
\description{
finding all observations to create a vector
}
