% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubjectsLabels.R
\name{getSubjectsLabels}
\alias{getSubjectsLabels}
\title{finding all subjects to create a vector}
\usage{
getSubjectsLabels(data, observation_column)
}
\arguments{
\item{data}{a data table normalized by the build_start_stop_table function}

\item{observation_column}{name of the observation column}
}
\value{
a vector which contain all the subjects observed
}
\description{
finding all subjects to create a vector
}
