% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDataTable.R
\name{importDataTable}
\alias{importDataTable}
\title{Get the data table from a file}
\usage{
importDataTable(inFile, header, sep, quote)
}
\arguments{
\item{inFile}{the file to import}

\item{header}{if the first line of the file is the header of the table}

\item{sep}{the separator used in the file}

\item{quote}{the quote used in the file}
}
\value{
the table table imported
}
\description{
Get the data table from a file
}
