% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildOnoTable.R
\name{buildOnoTable}
\alias{buildOnoTable}
\title{Transform data into a table of occurrences - non occurrences}
\usage{
buildOnoTable(df)
}
\arguments{
\item{df}{a \code{data.param} with a column of timestamps and a column of behaviors.
#'}
}
\value{
returns a \code{data.frame} of occurences - non occurences.
}
\description{
Transform data into a table of occurrences - non occurrences
}
